Changed motor representation as no longer needed
