+++ spynnaker_external_devices_plugin/pyNN/__init__.py
@@ -18,6 +18,8 @@ from spynnaker_external_devices_plugin.pyNN.external_devices_models.\
     pushbot_retina_device import PushBotRetinaResolution
 from spynnaker_external_devices_plugin.pyNN.external_devices_models.\
     pushbot_retina_device import PushBotRetinaPolarity
+from spynnaker_external_devices_plugin.pyNN.external_devices_models.\
+    munich_motor_device import MunichMotorDevice
 from spynnaker_external_devices_plugin.pyNN import model_binaries
 from spynnaker_external_devices_plugin.pyNN.\
     spynnaker_external_device_plugin_manager import \
@@ -143,72 +145,10 @@ def activate_live_output_to(population, device):
         population._get_vertex, device._get_vertex)
 
 
-def MunichMotorPopulation(
-        spinnaker_link_id, speed=30, sample_time=4096, update_time=512,
-        delay_time=5, delta_threshold=23, continue_if_not_different=True,
-        model=IF_curr_exp, params=None):
-    """ Create a population of 6 neurons which will drive the robot motor.
-        Neuron id 0 drives the forwards direction
-                  1 drives the backwards direction
-                  2 drives the left direction
-                  3 drives the right direction
-                  4 drives the clockwise direction
-                  5 drives the counter-clockwise direction
-        The control works by counting the number of spikes sent to each\
-        neuron, and then taking the difference between opposite\
-        directions at every sample time timesteps.  If the difference is\
-        above the threshold, a command is sent to move the robot in the\
-        appropriate direction.
-        The robot currently needs to be assigned a "virtual chip", which\
-        is then used to determine the routing keys that will be used for\
-        the motor control.  This chip must be one which does not exist\
-        within the current physical machine.
-
-    :param spinnaker_link_id: the id for the spinnaker link
-    :type spinnaker_link_id: int
-    :param speed: The speed to be sent to the motor when a direction is\
-                activated
-    :type speed: int
-    :param sample_time: The number of time steps between samples of the\
-                spike counts to determine the motor direction
-    :type sample_time: int
-    :param update_time: The number of time steps between sending updates\
-                of the motor speed
-    :type update_time: int
-    :param delay_time: The minimum number of microseconds between commands\
-                sent to the robot
-    :type delay_time: int
-    :param delta_threshold: The threshold of the difference between spike\
-                counts of opposite directions over which a direction will\
-                be chosen
-    :type delta_threshold: int
-    :param continue_if_not_different: Determines if the robot should\
-                continue in the current direction until a new direction\
-                is chosen, or if it should only move for one period\
-                before stopping unless the same direction is chosen\
-                again
-    :type continue_if_not_different: bool
-    :param model: The neuron model to use for the pre-motor population
-    :type model: class
-    :param params: Dictionary of parameters to give to the model
-    :type params: dict or None
-    :return: The newly created population
-    :rtype: Population
-    """
-    # checker for params and not having the setter for mutliple versions of
-    # munich motor control having cloned data
-    if params is None:
-        params = {}
-    return spynnaker_external_devices.create_munich_motor_population(
-        spinnaker_link_id, speed, sample_time, update_time, delay_time,
-        delta_threshold, continue_if_not_different, model, params)
-
-
 def SpikeInjector(
-        n_neurons, machine_time_step, timescale_factor, spikes_per_second,
-        ring_buffer_sigma, label, port, virtual_key=None,
-        database_notify_host=None, database_notify_port_num=None,
-        database_ack_port_num=None):
+        n_neurons, machine_time_step, timescale_factor, label, port,
+        virtual_key=None, database_notify_host=None,
+        database_notify_port_num=None, database_ack_port_num=None):
     """
     supports adding a spike injector to the applciation graph.
     :param n_neurons: the number of neurons the spike injector will emulate
@@ -218,12 +158,6 @@ def SpikeInjector(
     :param timescale_factor: the amount of scaling needed of the machine time
     step (basically a slow down function)
     :type timescale_factor: int
-    :param spikes_per_second: the expected number of spikes per second
-    :type spikes_per_second: int
-    :param ring_buffer_sigma: the number of standard divations from a
-    calcuation on how much safety in percision vs overflowing
-     (this is deduced from the front end)
-    :type ring_buffer_sigma: int
     :param label: the label given to the population
     :type label: str
     :param port: the port number used to listen for injections of spikes
@@ -262,6 +196,5 @@ def SpikeInjector(
     spynnaker_external_devices.add_socket_address(database_socket)
     return SpynnakerExternalDeviceSpikeInjector(
         n_neurons=n_neurons, machine_time_step=machine_time_step,
-        timescale_factor=timescale_factor, spikes_per_second=spikes_per_second,
-        ring_buffer_sigma=ring_buffer_sigma, label=label, port=port,
+        timescale_factor=timescale_factor, label=label, port=port,
         virtual_key=virtual_key)
